refactor pagefaulting
This is mostly a note to myself.

While I was thinking about the kernel, I realized that we can't just
'kill' a process outright, as it might have active locks etc. which
would cause a resource leak.

So then I realized that signals are actually an amazing way to cleanup.
When we return to userspace the kernel stack must be empty, and the only
resources that the process still has should be in the process struct. So
instead of stopping processes outright, the only way to stop a process
is through the exit syscall or through a signal, for which usually the
default handler is just a handler that calls exit. This makes sure that
the kernel releases any locks it has and frees temporary resources in a
clean way.

This sounded like a great idea, but there was one problem, uaccess. The
kernel needs to copy data from and to userspace memory and at the time,
the way this happened was through a special function in a special
section, which we could detect in the pagefault handler, if a pagefault
happened in this special section, it would go through the user pagefault
handler. Perhaps causing the process to be killed if it wasn't allowed
to access it.

I wasn't really sure how to apply this new signal/exit strategy to the
uaccess functions, as the function should return (if it didn't it would
likely cause a resource leak).

The uaccess functions may pagefault because of context switches, or
perhaps even the process itself unmapping the memory.

First I thought of perhaps "locking" the accessed pages in uaccess so
that they couldn't be unmapped, but this seemed really convoluted and
not a nice fix. And this doesn't fix the context switch problem. Also
locking for memory accesses just seems to be terribly slow anyway.

The latter could I guess be fixed by mapping it into kernel memory, but
that would again not be a pretty fix, it might not fit in high memory or
I had to place in below high memory and keep track of it. Not nice.

Then I thought of an atomic flag that would be set on an error in the
pagefault handler which uaccess has to read every single time. I guess a
better solution but still not really pretty.

Then I remembered reading about how linux does it, and something about
that the pagefault handler "fixes" the uaccess functions to return. At
the time of reading I found this very weird, I thought that when a
segfault happens, that we should kill the offending process ASAP. But
now I realize more what they were doing. Honestly I don't know if it
really is like this. But the solution seems ok to me.

So now (NOT TESTED) what the pagehandler does, is that when it finds
that the process was not allowed to do what it tried to do, it checks if
it was in a uaccess function, and if so makes the function return an
error value by doing some architecture magic.

Again these notes are mostly for myself