remove refcount from process
The refcount was needed to work on other processes that the process
itself, but after realizing that this doesn't really happen, I removed
it. A process can always safely use its own process handle, as a process
can only by terminated in sched_resume, or sched_do_kill which must
happen just before returning to userspace.